unified_platform/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-tobacco-erp-api"
version = "0.1.0"
description = "Unified Tobacco ERP backend (Phase 1 foundation)"
requires-python = ">=3.11"
dependencies = [
  "alembic==1.14.1",
  "celery==5.4.0",
  "email-validator==2.2.0",
  "fastapi==0.115.6",
  "passlib[bcrypt]==1.7.4",
  "psycopg[binary]==3.2.3",
  "pydantic==2.10.4",
  "pydantic-settings==2.7.1",
  "python-jose[cryptography]==3.3.0",
  "python-multipart==0.0.20",
  "redis==5.2.1",
  "sqlalchemy==2.0.36",
  "uvicorn[standard]==0.32.1",
]

[project.optional-dependencies]
dev = [
  "httpx==0.28.1",
  "pytest==8.3.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]
